fix: label and entry overlap in email field
The label of the email field used to return to its initial position
whenever the field losed focus and was invalid, which
sometimes could result in overlap with the text entried by the user.

Change the approach for the movement of labels, which used to rely on
input focus and validity, to the input class 'label-up', that provides
scripts full control over label positioning by class management.